% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{calculate.ttests}
\alias{calculate.ttests}
\title{Calcualte multiple t-tests}
\usage{
calculate.ttests(numericData = "", sampleInfo = "", cpdInfo = "",
  ttestgroups = "")
}
\arguments{
\item{numericData}{an metabolomics dataset}

\item{sampleInfo}{sample metadata file}

\item{cpdInfo}{compound data dictionary}

\item{ttestgroups}{ttest parameter file}
}
\value{
student t-test result in an excel sheet - Student_t_test_results.xlsx
}
\description{
Calcualte multiple t-tests
}
\examples{
calculate.ttests(numericData = "data_matrix.csv", sampleInfo ="sample_metadata.csv", cpdInfo="data_dictionary.csv", ttestgroups="ttest_param.csv")
}
